third_party/blink/web_tests/platform/generic/http/tests/devtools/runtime/runtime-es6-setSymbolPropertyValue-expected.txt

Tests editing Symbol properties.


Running: testSetUp
runtime-es6-setSymbolPropertyValue.js:15 ===== Initial =====
runtime-es6-setSymbolPropertyValue.js:16 2
runtime-es6-setSymbolPropertyValue.js:17 

Running: testSetSymbolPropertyValue
runtime-es6-setSymbolPropertyValue.js:15 ===== Set property =====
runtime-es6-setSymbolPropertyValue.js:16 3
runtime-es6-setSymbolPropertyValue.js:17 

Running: testDeleteSymbolProperty
runtime-es6-setSymbolPropertyValue.js:15 ===== Delete property =====
runtime-es6-setSymbolPropertyValue.js:16 undefined
runtime-es6-setSymbolPropertyValue.js:17 
